pyproject.toml
[tool.poetry]
name = "voice-agent-package"
version = "1.0.0"
description = "Voice agent functionality with Twilio and OpenAI integration"
authors = ["Mohd Mursaleen mohdmursaleen1207@gmail.com"]
readme = "README.md"
packages = [{include = "app"}]



[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.115.0"
uvicorn = "^0.27.0"
openai = "^1.68.0"
websockets = "^12.0"
twilio = "^8.13.0"
python-dotenv = "^1.0.0"
supabase = "^2.13.0"
python-multipart = "^0.0.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "voice_agent_package.main:main" 